+++ models/demos/gemma3/conftest.py
@@ -7,6 +7,7 @@ import os
 import pytest
 
 import ttnn
+from models.tt_transformers.demo.trace_region_config import get_supported_trace_region_size
 
 
 @pytest.fixture
@@ -19,6 +20,10 @@ def device_params(request, galaxy_type):
     )
     is_single_device = (mesh_device == (1, 1)) if isinstance(mesh_device, tuple) else (mesh_device == 1)
 
+    override_trace_region_size = get_supported_trace_region_size(request, mesh_device)
+
+    if override_trace_region_size:
+        params["trace_region_size"] = override_trace_region_size
     if "fabric_config" in params:
         if is_single_device:
             params["fabric_config"] = None
