+++ models/demos/gemma3/tt/model_config.py
@@ -1258,25 +1258,25 @@ class ModelArgs:
         self.trace_prefill_supported_seq_lens = self.get_trace_prefill_supported_seq_lens()
 
     def get_trace_prefill_supported_seq_lens(self):
+        default_supported_seq_lens = {
+            "N150": [128, 256, 512],
+            "N300": [128, 256, 512, 1024],
+            "T3K": [128, 256, 512, 1024, 2048, 4096, 8192],
+            "TG": [128, 256, 512, 1024, 2048, 4096, 8192],
+        }
+
         # TODO: If no specific sequence lengths are listed for a model and device, the default one will be used (from the default_supported_seq_lens dictionary)
-        supported_seq_lens = {
+        model_specific_supported_seq_lens = {
             # EXAMPLE: "gemma-3-4b": {
             #     "N150": [128, 256, 512, 1024, 2048],
             # }
-        }
-
-        default_supported_seq_lens = {
-            "N150": [128, 256, 512],
-            "N300": [128, 256, 512, 1024, 2048, 4096, 8192],
-            "T3K": [128, 256, 512, 1024, 2048, 4096, 8192],
-            "TG": [128, 256, 512, 1024, 2048, 4096, 8192],
         }
 
         model_name = self.base_model_name
         device_name = self.device_name
 
         # Try model-specific sequence lengths first
-        result = supported_seq_lens.get(model_name, {}).get(device_name)
+        result = model_specific_supported_seq_lens.get(model_name, {}).get(device_name)
         if result:
             return result
 
@@ -1287,9 +1287,6 @@ class ModelArgs:
 
         # No supported sequence lengths found, return empty list
         return []
-
-    def get_supported_trace_prefill_seq_lens(self):
-        return self.trace_prefill_supported_seq_lens
 
     def can_enable_trace(self, prefill_seq_len):
         """
@@ -1305,7 +1302,7 @@ class ModelArgs:
         if hasattr(self, "sliding_window") and getattr(self, "sliding_window") != None:
             return False
 
-        allowed_seq_lens = self.get_supported_trace_prefill_seq_lens()
+        allowed_seq_lens = self.trace_prefill_supported_seq_lens
 
         return (
             prefill_seq_len in allowed_seq_lens
